scripts/requirements.txt
nibabel==5.2.0
numpy==1.26.3
Pillow==10.2.0